requirements.txt
Django>=5.2.10
djangorestframework>=3.15.0
django-cors-headers>=4.5.0
Pillow>=11.0.0
django-filter>=24.3
python-decouple>=3.8
djangorestframework-simplejwt>=5.5.1
djangorestframework-simplejwt[crypto]
django-channels>=0.7.0
channels>=4.2.2
channels-redis>=4.3.0
dj-rest-auth>=7.0.2
djoser>=2.3.3
redis>=7.1.0
django-allauth>=0.61.0
openai>=1.0.0
psycopg2-binary>=2.9.9
daphne>=4.0.0
gunicorn>=21.2.0
pywebpush>=2.0.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.11.0
async-timeout==5.0.1
attrs==25.4.0
autobahn==25.12.2
Automat==25.4.16
cbor2==5.8.0
certifi==2026.1.4
cffi==2.0.0
channels==4.3.2
channels_redis==4.3.0
charset-normalizer==3.4.4
colorama==0.4.6
constantly==23.10.4
cryptography==46.0.3
daphne==4.2.1
defusedxml==0.7.1
distro==1.9.0
dj-database-url==3.1.0
dj-rest-auth==7.0.2
Django==5.2.10
django-allauth==65.13.1
django-channels==0.7.0
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
djoser==2.3.3
frozenlist==1.8.0
gunicorn==23.0.0
h11==0.16.0
http-ece==1.2.1
httpcore==1.0.9
httpx==0.28.1
hyperlink==21.0.0
idna==3.11
Incremental==24.11.0
jiter==0.12.0
msgpack==1.1.2
multidict==6.7.0
oauthlib==3.3.1
openai==2.15.0
packaging==25.0
pillow==12.1.0
propcache==0.4.1
psycopg2-binary==2.9.11
py-ubjson==0.16.1
py-vapid==1.9.4
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.10.1
pyOpenSSL==25.3.0
python-decouple==3.8
python3-openid==3.2.0
pywebpush==2.2.0
redis==7.1.0
requests==2.32.5
requests-oauthlib==2.0.0
service-identity==24.2.0
six==1.17.0
sniffio==1.3.1
social-auth-app-django==5.7.0
social-auth-core==4.8.3
sqlparse==0.5.5
tqdm==4.67.1
Twisted==25.5.0
txaio==25.12.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
ujson==5.11.0
urllib3==2.6.3
yarl==1.22.0
zope.interface==8.2
